+++ src/components/Main/WholeApartInfo/ServiceNotice/ServiceNoticePagination.module.scss
@@ -1,0 +1,24 @@
+@use "@/styles/colors.scss";
+
+.pagination {
+  margin-top: auto;
+  display: flex;
+  justify-content: center;
+  gap: 0.4rem;
+  padding-top: 1.6rem;
+  .pageContainer {
+    display: flex;
+    justify-content: center;
+    gap: 0.8rem;
+    .pageLink {
+      font-size: 1.4rem;
+      font-weight: 500;
+      line-height: 2rem;
+      color: colors.$mainTitle;
+    }
+    .pageActive {
+      font-weight: 700;
+      color: colors.$mainRed;
+    }
+  }
+}
